pyproject.toml
[project]
name = "grandmaster-mcp"
version = "0.1.0"
description = "Chess intelligence layer using Python and MCP"
authors = [
    { name = "Antigravity", email = "antigravity@google.com" }
]
dependencies = [
    "mcp[cli]>=1.3.0",
    "python-chess>=1.10.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "fastmcp>=0.4.1",
]
requires-python = ">=3.12"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["frontend*", ".venv*", "reference_mcp*"]
